backend/requirements.txt
asgiref==3.7.2
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
filelock==3.12.3
Flask==2.3.3
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.23.5
pandas==1.5.3
torch==2.1.2
torchvision==0.2.0
transformers==4.36.2
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.13
Werkzeug==2.3.7

